Added_Error_measure
It's for when sending an invalid character like an alphabet which outputs a positive response, I added a measure that prints positive response in capitalized format to alarm for presence of an invalid input incase the variable "no" is used elsewhere. Add an assertion function to stop incase having a non integer input is a problem.